Fix various warnings (both implicit conversions and real bugs).
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3110 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/Terrain.h
@@ -62,12 +62,12 @@ public:
 	// calculate the vertex under a given position (rounding down coordinates)
 	static void CalcFromPosition(const CVector3D& pos, i32& i, i32& j)
 	{
-		i = pos.X / CELL_SIZE; j = pos.Z / CELL_SIZE;
+		i = (i32)(pos.X / CELL_SIZE); j = (i32)(pos.Z / CELL_SIZE);
 	}
 	// calculate the vertex under a given position (rounding down coordinates)
 	static void CalcFromPosition(float x, float y, i32& i, i32& j)
 	{
-		i = x / CELL_SIZE; j = y / CELL_SIZE;
+		i = (i32)(x / CELL_SIZE); j = (i32)(y / CELL_SIZE);
 	}
 	// calculate the normal at a given vertex
 	void CalcNormal(u32 i, u32 j, CVector3D& normal) const;
